pyproject.toml
[project]
name = "lubw-client"
version = "0.1.0"
description = "client to call sensor data from Landesamt für Umwelt Baden Würrtemberg (LUBW)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler>=3.11.1",
    "numpy>=2.2.6",
    "paho-mqtt>=2.1.0",
    "pandas>=2.3.3",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
